Move weaviate examples to examples folder
