+++ examples/weaviate/embedding_biology.py
