+++ examples/weaviate/text_search_local_embedder.py
