+++ examples/weaviate/text_search_remote_embedder.py
