+++ examples/weaviate/vector_search.py
@@ -16,13 +16,7 @@
 from neo4j_genai.retrievers.external.weaviate import WeaviateNeo4jRetriever
 from neo4j import GraphDatabase
 import weaviate
-import sys
-import os
-
-BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-sys.path.append(f"{BASE_DIR}/../../../../../../tests/e2e/weaviate_e2e")
-
-from utils import get_query_vector  # noqa
+from embedding_biology import EMBEDDING_BIOLOGY
 
 NEO4J_URL = "neo4j://localhost:7687"
 NEO4J_AUTH = ("neo4j", "password")
@@ -38,7 +32,7 @@ def main():
                 id_property_external="neo4j_id",
                 id_property_neo4j="id",
             )
-            res = retriever.search(query_vector=get_query_vector(), top_k=2)
+            res = retriever.search(query_vector=EMBEDDING_BIOLOGY, top_k=2)
             print(res)
